[minor] new conversion helpers
A common use case for certificates is with tls clients and servers.
These functions basically always consume a `tls.Config` struct with the
`Certificates` and/or `RootCAs` fields populated.

Add two new methods to the `KeyAndCert` struct to export both types and
make the return signature match the expectations of their relevant
fields in `tls.Config`:

```go
tls.Config{
  Certificates: k.TLSCertificate(), // tls.Certificate
  RootCAs: k.CertPool(),            // *x509.CertPool
}
```